Successfully tested Interview Updates

+++ interviewbit/app/controllers/interviews_controller.rb
@@ -22,16 +22,14 @@ class InterviewsController < ApplicationController
     @users=User.find(params[:interview][:users])
     @i=Interview.new(i_para)
     @i.users = @users
-    puts @i.errors.full_messages
-    if @i.save
+    if @i.save(context: :create)
       @u=User.joins(:interviews).where("interviews_users.interview_id = ?", @i.id)
       arg = Array.new
       arg.append(@i.id)
       @u.each do |u|
         arg.append(u.id)
       end
-      #UserMailer.with(arg: arg).new_interview_email.deliver_later
-      UserMailer.new_interview_email.perform_async(arg)
+      UserMailer.with(arg: arg).new_interview_email.deliver_later
       redirect_to '/'
     else
         render 'new'
@@ -42,26 +40,53 @@ class InterviewsController < ApplicationController
     params[:interview][:start]=params[:interview][:start]+" "+params[:interview][:start_time]
     params[:interview][:finish]=params[:interview][:finish]+" "+params[:interview][:finish_time]
     params[:interview][:users].shift
-    users=User.find(params[:interview][:users])
     @i = Interview.find(params[:id])
-    id=@i.id
-    query = "delete from interviews_users where interview_id = "+String(id)
-    ActiveRecord::Base.connection.execute(query)
-    @i.users = users
-    if @i.update(i_para)
-      arg =Array.new
-      arg.append(@i.title)
-      arg.append(@i.start)
-      arg.append(@i.finish)
-      users.each do |u|
-        arg.append(u.email)
-      end
-      UserMailer.with(arg: arg).update_interview_email.deliver_later
-
-
-      redirect_to interviews_path
+    if params[:interview][:users].size < 2
+      flash.now[:notice] = "Select atleast 2 Users."
+      render 'edit' 
     else
-      render 'edit'
+      flag=false
+      cur_usr=Array.new
+        @i.users.each do |cu|
+          cur_usr.append(cu.id)
+      end
+      query = "delete from interviews_users where interview_id = "+String(@i.id)
+      ActiveRecord::Base.connection.execute(query)  
+      cnt=1
+      params[:interview][:users].each do |sel_usr|
+        if check(sel_usr,params[:interview][:start],params[:interview][:finish]) ==false
+          uid=sel_usr
+          flash.now[cnt]= String(User.find(uid).name)+" is unavailable in this time slot"
+          cnt=cnt+1
+          flag=true
+        end
+      end
+      if flag==true
+        users=User.find(cur_usr)
+        @i.users << users
+        render 'edit'
+      else
+        users=User.find(params[:interview][:users])
+        id=@i.id
+        if @i.update(i_para)
+          @i.users << users
+          arg =Array.new
+          arg.append(@i.title)
+          arg.append(@i.start)
+          arg.append(@i.finish)
+          users.each do |u|
+            arg.append(u.email)
+          end
+          UserMailer.with(arg: arg).update_interview_email.deliver_later
+          redirect_to interviews_path
+        else
+          query = "delete from interviews_users where interview_id = "+String(id)
+          ActiveRecord::Base.connection.execute(query)
+          users=User.find(cur_usr)
+          @i.users << users
+          render 'edit'
+        end
+      end
     end
   end
 
@@ -83,4 +108,19 @@ class InterviewsController < ApplicationController
     def i_para
         params.require(:interview).permit(:title,:start,:finish)
     end 
+    def check(uid,start,finish)
+      start=start + " UTC"
+      finish=finish + " UTC"
+      usr=User.find(uid)
+      ans = true
+      inter = Interview.joins(:users).where("interviews_users.user_id = ?", uid )
+        inter.each do |i|
+          st=i.start
+          fs=i.finish
+          if ((st>=start && st<=finish) || (fs>=start && fs<=finish) || (st<=start && fs>=finish)) ==true
+            ans = false 
+          end
+        end
+        return ans
+    end
 end
